Refactor plotter and use xesmf for regridding
